requirements.txt
groq>=0.9.0
duckduckgo-search>=5.3.1b1
python-dotenv>=1.0.1
rich>=13.7.1
